Create hook user list
