+++ src/pages/HomePage.tsx
@@ -1,43 +1,20 @@
-import { useState } from "react";
-import { User } from "../interfaces";
-import { getUserFollowers, getUserList } from "../services";
+
 import { Background, UserTable, Input, NotificationManager, GraphBar } from "../components";
+import { useUserList } from "../hooks/useUserList";
 
 export const HomePage = () => {
 
-    const [query, setQuery] = useState('');
-    const [users, setUsers] = useState<User[]>([]);
-    const [emptyUser, setEmptyUser] = useState('');
-    const [error, setError] = useState<string | null>(null);
-    const [followersData, setFollowersData] = useState<number[]>([]);
-    const [userNames, setUserNames] = useState<string[]>([]);
-
-    const handleSearch = () => {
-        if (query.length < 4 || query === 'iseijasunow') {
-            setError('El término de búsqueda no es válido');
-            setUsers([]);
-            setQuery('');
-            return;
-        }
-        getUserList(query)
-            .then(async (users) => {
-                const slicedUsers = users.slice(0, 10);
-                setUsers(slicedUsers);
-                const followersPromises = slicedUsers.map(user => getUserFollowers(user.login));
-                setUserNames(slicedUsers.map(user => user.login));
-                const followersCounts = await Promise.all(followersPromises);
-                setFollowersData(followersCounts);
-                users.length === 0 && setEmptyUser('No user found, try again....');
-            })
-            .catch(error => {
-                console.log(error);
-                setError('Error getting users')
-            });
-    };
-
-    const handleClose = () => {
-        setError(null);
-    };
+    const {
+        error,
+        query,
+        users,
+        followersData,
+        userNames,
+        emptyUser,
+        handleClose,
+        handleSearch,
+        setQuery
+    } = useUserList();
 
     return (
         <>
